// src/components/IconButton/IconButton.types.ts
// Generated with util/create-component.js
import { ButtonProps } from 'react-bootstrap/Button'

export interface IconButtonProps extends ButtonProps {
  icon:
    | 'plusCircle'
    | 'close'
    | 'back'
    | 'forward'
    | 'up'
    | 'down'
    | 'refresh'
    | 'add'
    | 'edit'
    | 'delete'
    | 'remove'
    | 'save'
    | 'spinner'
    | 'menu'
    | 'check'
    | 'signout'
  variant?: string
  disabled?: boolean
  spin?: boolean
  className?: string
  title?: string
}